+++ src/UIComponents/molecules/ProfileAvatar/index.js
@@ -1,0 +1,2 @@
+import comp from "./ProfileAvatar";
+export default comp;